feat(vmop): first implementation (#4)
## Description
 Implementation resource VirtualMachineOperation. 
The VirtualMachineOperation custom resource allows for declarative
management of a working virtual machine. It enables actions such as
stopping, starting, and restarting the virtual machine.

## Why do we need it, and what problem does it solve?
The VirtualMachineOperation resource is needed to provide a high-level
abstraction for managing virtual machines. It solves the problem of
efficiently and easily controlling the state of virtual machines through
a declarative approach, rather than relying on manual, time-consuming
operations.

## What is the expected result?
VirtualMachineOperation is an immutable resource. After deployment, it
can have 4 phases of status: Completed, Failed, InProgress, Pending.

---------

Signed-off-by: Yaroslav Borbat <yaroslav.borbat@flant.com>
Signed-off-by: Yaroslav Borbat <86148689+yaroslavborbat@users.noreply.github.com>
Co-authored-by: Ivan Mikheykin <ivan.mikheykin@flant.com>