fix(vd): optimize the creation time for empty disks (#786)
Blank disk could remain in the 'Failed' status, even though its PVC was ready for use.
The issue was on the CDI side. As part of this PR, the creation of blank disks was refactored and switched to a step-based design.
This allowed to avoid the error in CDI, sped up the disk creation process, and simplified the maintenance of blank disks.

The StorageClassReady handler was refactored to simplify its maintenance.



---------

Signed-off-by: Isteb4k <dmitry.rakitin@flant.com>