feat(vm): calculate sockets count by cores count (#747)
## Description
Update CPU configuration to dynamically calculate the number of sockets based on the number of cores and set a maximum of 8 sockets.

## Why do we need it, and what problem does it solve?
Currently, each core has its own socket. However, some operating systems do not function correctly with this setup.

## What is the expected result?
The system will dynamically calculate the number of sockets based on the number of cores, ensuring compatibility with various operating systems. The maximum number of sockets will be limited to 8.

---------

Signed-off-by: Daniil Antoshin <daniil.antoshin@flant.com>
Signed-off-by: mortelumina98 <ekaterina.matseevska@flant.com>
Signed-off-by: Ekaterina Matseevska <ekaterina.matseevska@flant.com>
Co-authored-by: Pavel Tishkov <pavel.tishkov@flant.com>
Co-authored-by: mortelumina98 <ekaterina.matseevska@flant.com>