fix(vm): ensure detachment happened (#841)
BlockDeviceHandler was violating the contract regarding the setting of .status.blockDeviceRefs.
BlockDeviceHandler was populating .status.blockDeviceRefs based on what was in .spec and from the existing vmbda in the cluster, which led to problems.
.status.blockDeviceRefs should reflect not only the block devices expected to be attached as per the spec but also those that are present in kvvm (these two sets of block devices may not match). Fixed this.

The log level of conflict error (“the object has been modified; please apply your changes to the latest version and try again”) for virtual machines is now set to debug, not error.

The vmbda will no longer be deleted until detaching is completed.

Signed-off-by: Isteb4k <dmitry.rakitin@flant.com>