test(vm): reboot using nohup (#901)
We had tests failing because the command ssh `reboot -f` was not executing completely. More precisely, the reboot request was sent, and the virtual machine rebooted successfully, but the SSH session remained active throughout the reboot process and eventually disconnected with a timeout error:

> Error: websocket: close 1006 (abnormal closure): unexpected EOF

This caused the test to fail. The `-f` flag was causing this behavior. Without it, everything worked fine. I resolved it by rebooting with `nohup reboot -f`.

Signed-off-by: Isteb4k <dmitry.rakitin@flant.com>