test: increase e2e tests timeout (#914)
Description
This change increases the timeout duration for Ginkgo-based end-to-end (e2e) tests.

Why is this change necessary, and what problem does it address?
As the number of e2e tests has grown significantly, the current timeout of 1 hour is no longer sufficient to complete all test cases. This limitation has resulted in incomplete test runs and potential disruptions to the testing process. By extending the timeout, we ensure that all tests have adequate time to execute, regardless of their increasing complexity and count.

What is the expected outcome?
The timeout for all e2e tests will be increased to 2 hours . This adjustment will allow the entire test suite to run to completion without premature termination, ensuring reliable and consistent test results.

---------

Signed-off-by: Daniil Antoshin <daniil.antoshin@flant.com>