feat(vm):  set error policy report as default (#983)
Change VM behavior to I/O problems for disks.

Previously, the presence of I/O problems stopped the VM and paused it.

Now, instead of stopping the VM, I/O errors will be reported to the guest OS, allowing the guest system to deal with the problem (e.g., through retry mechanisms,
failover).

Signed-off-by: Yaroslav Borbat <yaroslav.borbat@flant.com>