feat(ci): add alias rules for golangci-lint (#1387)
Description
This PR adds a new linter for golangci-lint. This linter is responsible for import aliases.

Why do we need it, and what problem does it solve?
First and foremost, to maintain consistency. There is a list of aliases that we use always and everywhere, so it would be wise to check their correctness with a linter and protect developers from typos and errors.

What is the expected result?
New linter for all golangci-lint rulesets.

---------

Signed-off-by: Daniil Antoshin <daniil.antoshin@flant.com>
(cherry picked from commit 21b3b137f697342caec9838f05769a70df00187e)
Signed-off-by: Isteb4k <dmitry.rakitin@flant.com>