fix(vm): recreate unschedulable intvirtvmi's pod (#1512)
This workaround is required due to a bug in the KVVM workflow.
When a KVVM is created with conflicting placement rules and cannot be scheduled,
it remains unschedulable even if these rules are changed or removed.

Signed-off-by: Roman Sysoev <roman.sysoev@flant.com>
(cherry picked from commit 69d4f1a8caf6d784f154f85e0ded045b882b168c)
Signed-off-by: Isteb4k <dmitry.rakitin@flant.com>