fix(vmbda): report error when device is unavailable on VM node (#1561)
Add condition error for PersistentVolume node affinity when attached block devices via VMBDA is not available on VM node.

When using storage classes with node affinity, the PersistentVolume is pinned to a specific node. If a user tries to hot-plug such a disk to a VM running on a different node, the attachment fails silently or with unclear errors.

Signed-off-by: Daniil Loktev <lokt.daniil@gmail.com>