fix(vdsnapshot): resolve fsfreeze race condition (#1668)
Sometimes, when the virtual disk snapshot controller creates a batch of snapshots with required consistency, a race condition in the freeze or unfreeze "filesystem" operation can cause an inconsistent snapshot. To prevent this, the KubeVirt virtual machine instance is now annotated with a "freeze" or "unfreeze" request, and the service checks that the type of this request matches the "fsFreezeStatus". The "fsFreezeStatus" is considered trusted only if it matches the request type.

Signed-off-by: Roman Sysoev <roman.sysoev@flant.com>
Co-authored-by: Isteb4k <dmitry.rakitin@flant.com>